pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rai-checklist-cli"
version = "0.7.3"  
description = "A CLI tool to generate responsible AI checklists for machine learning projects."
readme = "README.md"
authors = [{name = "Noble Ackerson", email = "noblel@byteanatom.com"}]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.6"
dependencies = [
    "pyyaml",
    "tqdm",
    "ipywidgets",
    "google-generativeai",  # Added new dependency
]

[project.optional-dependencies]
test = ["pytest", "pylint"]  # Moved 'pylint' to test dependencies

[project.scripts]
rai-checklist = "rai_checklist_cli.cli:main"

[tool.setuptools.packages.find]
include = ["rai_checklist_cli*"]

[tool.setuptools.package-data]
rai_checklist_cli = ["templates.yaml"]
